4A - Watermelon: Refactor the Python solution
First, we use sys.stdin.readline() for faster input processing. Such
a change, most likely, will not make any mesurable difference in the
problem at hand; it is considered good practice, however, to keep the
coding style consistent. So, in anticipation of these huge inputs, we
employ sys module's faster input methods.

Second, we reorder the "if" statement's expressions to remove
unnecessary comparsions for odd numbers greater than 3. This brings
the Python solution in line with the C++ one.